requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.4.0

# Vector DB & Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# PDF Processing
pypdf>=3.17.0
pymupdf>=1.23.0

# ML & NLP
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Retrieval
rank-bm25>=0.2.2

# LLM Clients
openai>=1.3.0
requests>=2.31.0

# Utilities
tqdm>=4.66.0
tiktoken>=0.5.1

# Testing
pytest>=7.4.0

# Optional: For generating sample PDFs (not required for normal use)
# reportlab>=4.0.0

